requirements/development.txt
-r base.txt
black==22.3.0
flake8>=3.6.0
isort==5.6.4  # leave this pinned - it tends to change rules between patch releases
flake8-blind-except==0.1.1
flake8-comprehensions==3.8.0
flake8-print==5.0.0
flake8-assertive==2.0.0
curlylint==0.13.1
djhtml==1.4.13
pytest-playwright>=0.3.3,<1
